requirements.txt
seaborn
imblearn.over_sampling
matplotlib
sklearn
tensorflow